Move some easy stuff into runners/core-java
This change moves a set of classes with no dependents,
leaving them in the same Java packages for now.
