+++ runners/core-java/src/main/java/org/apache/beam/sdk/util/SystemReduceFn.java
