+++ runners/core-java/src/test/java/org/apache/beam/sdk/util/ReduceFnTester.java
